Move YAML-parsing code into vectordatasource library, so we can re-use it to add a test ensuring we don't backslide on kind & kind_detail whitelists.
